feat: update ORM models and seed logic, add user progress tracking

+++ backend/internal/models/cafe_order.go
@@ -5,6 +5,6 @@ type CafeOrder struct {
 	ID         uint64 `json:"id" gorm:"primaryKey;autoIncrement"`
 	Name       string `json:"name" gorm:"not null"`
 	Category   string `json:"category"` // e.g., "Coffee", "Snack", "Meal"
-	EnergyCost int    `json:"energy_cost" gorm:"not null"`
-	RewardXP   int    `json:"reward_xp" gorm:"not null"`
+	EnergyCost int64  `json:"energy_cost" gorm:"not null"`
+	RewardXP   int64  `json:"reward_xp" gorm:"not null"`
 }
